Reduce size of filter dac
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/libsidplayfp@1498 75748be7-0387-4326-94c7-f41ce0883d2c

+++ builders/residfp-builder/residfp/FilterModelConfig.h
@@ -127,7 +127,7 @@ public:
      * @param adjustment
      * @return the DAC table
      */
-    unsigned int* getDAC(double adjustment) const;
+    unsigned short* getDAC(double adjustment) const;
 
     Integrator* buildIntegrator();
 };
